refactor: Move exception handler filter
Common should be for stuff that gets imported by various modules, this doesn't seem like it belongs in there
